finalize file structure
