+++ seleTest.py
@@ -175,8 +175,6 @@ class LoginLogoutRegisterTest(unittest.TestCase):
 		self.driver.close()
 
 
-
-
 if __name__ == "__main__":
     unittest.main()
