+++ survey/templates/base.html
@@ -32,18 +32,7 @@
 					<a class="nav-link" href="{{ url_for('main.index') }}">Home <span class="sr-only">(current)</span></a>
 				</li>
 				<li class="nav-item">
-					<a class="nav-link" href="#">Link</a>
-				</li>
-				<li class="nav-item dropdown">
-					<a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-						Polls
-					</a>
-					<div class="dropdown-menu" aria-labelledby="navbarDropdown">
-						<a class="dropdown-item" href="#">See result</a>
-						<a class="dropdown-item" href="{{ url_for('polls.active_polls') }}">Active poll</a>
-						<div class="dropdown-divider"></div>
-						<a class="dropdown-item" href="#">Something else here</a>
-					</div>
+					<a class="nav-link" href="{{ url_for('polls.active_polls') }}">Active poll</a>
 				</li>
 				<li class="nav-item">
 					{% if current_user.is_admin %}
