+++ survey/templates/user-admin/category.html
@@ -1,4 +1,7 @@
-<p id="Category_description">You are able to add and delete Categories!</p>
+<div class="row">
+	<p id="Category_description">You are able to add and delete Categories!</p>
+	<button type="button" id="poll_add" class="btn btn-light col-xs-2 admin_add_button" data-toggle="modal" data-target="#add_category"><i class="fa fa-plus"></i></button>
+</div>
 <table id="category_table" class="transparent">
 	<thead>
 		<tr>
