+++ survey/templates/user-admin/users.html
@@ -1,4 +1,7 @@
-<p id="user_description">You are able to add and delete users!</p>
+<div class="row">
+	<p id="user_description">You are able to add and delete users!</p>
+	<button type="button" id="poll_add" class="btn btn-light col-xs-2 admin_add_button" data-toggle="modal" data-target="#add_user"><i class="fa fa-plus"></i></button>
+</div>
 <table id="user_table" class="transparent">
 	<thead>
 		<tr>
