+++ survey/templates/user/reset_password.html
