+++ survey/templates/user/reset_token.html
